+++ packages/widget/src/components/Pages/Swap/Withdraw/Wallet/index.tsx
@@ -14,6 +14,7 @@ import { ActionMessages } from "../messages/TransactionMessages";
 import { useTransfer } from "@/hooks/useTransfer";
 import { useRpcHealth } from "@/context/rpcHealthContext";
 import RPCUnhealthyMessage from "./RPCUnhealthyMessage";
+import SwapError from "@/components/Pages/Swap/Form/SecondaryComponents/SwapError";
 
 type Props = {
     swapData: SwapBasicData
@@ -222,7 +223,8 @@ const TransferTokenButton: FC<TransferTokenButtonProps> = ({
         />
     }
 
-    return <div className="w-full space-y-3 flex flex-col justify-between h-full text-primary-text">
+    return <div className="w-full space-y-2 flex flex-col justify-between h-full text-primary-text">
+        <SwapError />
         {
             buttonClicked &&
             <ActionMessage