+++ packages/widget/src/types/addressUtils.ts
@@ -1,12 +1,1 @@
-export interface AddressUtilsProvider {
-    providerName: string,
-    supportsNetwork(network: { name: string }): boolean,
-    isValidAddress: (props: AddressUtilsProviderProps) => boolean,
-    addressFormat?: (props: AddressUtilsProviderProps) => string,
-}
-
-export type AddressUtilsProviderProps = {
-    address: string,
-    network?: { name: string } | null,
-    providerName?: string
-}
+export type { AddressUtilsProvider, AddressUtilsProviderProps } from "@layerswap/utils";