+++ packages/wallets/imtblPassport/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @layerswap/wallet-imtbl-passport
 
+## 1.4.0
+
+### Minor Changes
+
+- Add Deposit widget
+
+### Patch Changes
+
+- Updated dependencies
+  - @layerswap/widget@1.4.0
+
 ## 1.2.0
 
 ### Patch Changes