+++ packages/wallets/paradex/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @layerswap/wallet-paradex
 
+## 1.4.0
+
+### Minor Changes
+
+- Add Deposit widget
+
+### Patch Changes
+
+- Updated dependencies
+  - @layerswap/wallet-evm@1.4.0
+  - @layerswap/wallet-starknet@1.4.0
+  - @layerswap/widget@1.4.0
+
 ## 1.2.0
 
 ### Patch Changes